Changed and extended the groups menu pages
