+++ webapp/src/App.js
@@ -7,8 +7,7 @@ import Footer from './components/Footer';
 import Home from './pages/Home';
 import Homepage from './pages/Homepage';
 import Game from './pages/Game';
-import GroupList from './pages/GroupList';
-import GroupCreate from './pages/GroupCreate';
+import Groups from './pages/Groups';
 import {Route, Routes} from 'react-router-dom';
 import { createTheme, ThemeProvider } from '@mui/material/styles';
 import { Box } from '@mui/material';
@@ -43,8 +42,7 @@ function App() {
             <Route path="/instructions" element={<Instructions />}/>
             <Route path="/homepage" element={<Homepage />}/>
             <Route path="/game" element={<Game />}/>
-            <Route path="/group/list" element={<GroupList />}/>
-            <Route path="/group/create" element={<GroupCreate />}/>
+            <Route path="/group/menu" element={<Groups />}/>
           </Routes>
         <Footer/>
       </ThemeProvider>
