+++ webapp/src/components/NavBar.js
@@ -13,8 +13,7 @@ const pages = [
   { path: '/homepage', text: 'Play' },
   { path: '/statistics', text: 'Statistics' },
   { path: '/instructions', text: 'Instructions' },
-  { path: '/group/list', text: 'List Groups' },
-  { path: '/group/create', text: 'Create Group' }
+  { path: '/group/menu', text: 'Groups' }
   // Add an object for each new page
 ];
